Support float type inputs

+++ calculator/main.py
@@ -26,11 +26,11 @@ operations = {
     "/": divide,
 }
 exit_calculator = False
-num1 = int(input("What's the first number? "))
+num1 = float(input("What's the first number? "))
 display(operations)
 while not exit_calculator:
     operation = input("Pick an operation: ")
-    num2 = int(input("What's the next number? "))
+    num2 = float(input("What's the next number? "))
     result = operations[operation](num1, num2)
     print(f"{num1} {operation} {num2} = {result}")
     another_calculation = input(f"Type 'y' to continue calculating with {result}, or type 'n' to exit: ").lower()
